Added a Utils class
Added missing getCode methods
Added missing JavaDoc
Fixed test for CurrentGame

+++ src/main/java/no/stelar7/api/l4j8/basic/constants/GameQueueType.java
@@ -7,153 +7,153 @@ public enum GameQueueType
     /**
      * Custom games
      */
-    CUSTOM(0),
+    CUSTOM(0L),
     /**
      * Normal 3v3 games
      */
-    NORMAL_3x3(8),
+    NORMAL_3x3(8L),
     /**
      * Normal 5v5 Blind Pick games
      */
-    NORMAL_5x5_BLIND(2),
+    NORMAL_5x5_BLIND(2L),
     /**
      * Normal 5v5 Draft Pick games
      */
-    NORMAL_5x5_DRAFT(14),
+    NORMAL_5x5_DRAFT(14L),
     /**
      * Ranked Solo 5v5 games
      */
-    RANKED_SOLO_5x5(4),
+    RANKED_SOLO_5x5(4L),
     /**
      * Ranked Premade 5v5 games
      */
-    @Deprecated RANKED_PREMADE_5x5(6),
+    @Deprecated RANKED_PREMADE_5x5(6L),
     /**
      * Ranked Premade 3v3 games
      */
-    @Deprecated RANKED_PREMADE_3x3(9),
+    @Deprecated RANKED_PREMADE_3x3(9L),
     /**
      * Ranked Team 3v3 games
      */
-    RANKED_TEAM_3x3(41),
+    RANKED_TEAM_3x3(41L),
     /**
      * Ranked Team 5v5 games
      */
-    RANKED_TEAM_5x5(42),
+    RANKED_TEAM_5x5(42L),
     /**
      * Dominion 5v5 Blind Pick games
      */
-    ODIN_5x5_BLIND(16),
+    ODIN_5x5_BLIND(16L),
     /**
      * Dominion 5v5 Draft Pick games
      */
-    ODIN_5x5_DRAFT(17),
+    ODIN_5x5_DRAFT(17L),
     /**
      * Historical Summoner's Rift Coop vs AI games
      */
-    @Deprecated BOT_5x5(7),
+    @Deprecated BOT_5x5(7L),
     /**
      * Dominion Coop vs AI games
      */
-    BOT_ODIN_5x5(25),
+    BOT_ODIN_5x5(25L),
     /**
      * Summoner's Rift Coop vs AI Intro Bot games
      */
-    BOT_5x5_INTRO(31),
+    BOT_5x5_INTRO(31L),
     /**
      * Summoner's Rift Coop vs AI Beginner Bot games
      */
-    BOT_5x5_BEGINNER(32),
+    BOT_5x5_BEGINNER(32L),
     /**
      * Historical Summoner's Rift Coop vs AI Intermediate Bot games
      */
-    BOT_5x5_INTERMEDIATE(33),
+    BOT_5x5_INTERMEDIATE(33L),
     /**
      * Twisted Treeline Coop vs AI games
      */
-    BOT_TT_3x3(52),
+    BOT_TT_3x3(52L),
     /**
      * Team Builder games
      */
-    GROUP_FINDER_5x5(61),
+    GROUP_FINDER_5x5(61L),
     /**
      * ARAM games
      */
-    ARAM_5x5(65),
+    ARAM_5x5(65L),
     /**
      * One for All games
      */
-    ONEFORALL_5x5(70),
+    ONEFORALL_5x5(70L),
     /**
      * Snowdown Showdown 1v1 games
      */
-    FIRSTBLOOD_1x1(72),
+    FIRSTBLOOD_1x1(72L),
     /**
      * Snowdown Showdown 2v2 games
      */
-    FIRSTBLOOD_2x2(73),
+    FIRSTBLOOD_2x2(73L),
     /**
      * Summoner's Rift 6x6 Hexakill games
      */
-    SR_6x6(75),
+    SR_6x6(75L),
     /**
      * Ultra Rapid Fire games
      */
-    URF_5x5(76),
+    URF_5x5(76L),
     /**
      * Ultra Rapid Fire games played against AI games
      */
-    BOT_URF_5x5(83),
+    BOT_URF_5x5(83L),
     /**
      * Doom Bots Rank 1 games
      */
-    NIGHTMARE_BOT_5x5_RANK1(91),
+    NIGHTMARE_BOT_5x5_RANK1(91L),
     /**
      * Doom Bots Rank 2 games
      */
-    NIGHTMARE_BOT_5x5_RANK2(92),
+    NIGHTMARE_BOT_5x5_RANK2(92L),
     /**
      * Doom Bots Rank 5 games
      */
-    NIGHTMARE_BOT_5x5_RANK5(93),
+    NIGHTMARE_BOT_5x5_RANK5(93L),
     /**
      * Ascension games
      */
-    ASCENSION_5x5(96),
+    ASCENSION_5x5(96L),
     /**
      * Twisted Treeline 6x6 Hexakill games
      */
-    HEXAKILL(98),
+    HEXAKILL(98L),
     /**
      * Butcher's Bridge games
      */
-    BILGEWATER_ARAM_5x5(100),
+    BILGEWATER_ARAM_5x5(100L),
     /**
      * King Poro games
      */
-    KING_PORO_5x5(300),
+    KING_PORO_5x5(300L),
     /**
      * Nemesis games
      */
-    COUNTER_PICK(310),
+    COUNTER_PICK(310L),
     /**
      * Black Market Brawlers games
      */
-    BILGEWATER_5x5(313);
+    BILGEWATER_5x5(313L);
 
     public static GameQueueType getFromCode(final Number gameQueueConfigId)
     {
         return Stream.of(GameQueueType.values()).filter(t -> t.code == gameQueueConfigId).findFirst().get();
     }
 
-    Integer code;
+    Long code;
 
-    GameQueueType(final int code)
+    GameQueueType(final Long code)
     {
         this.code = code;
     }
 
-    public Integer getCode()
+    public Long getCode()
     {
         return this.code;
     }
